modutil: add support for loading some edge case ST modules.
Some old Soundtracker modules do NOT store patterns referenced after
the end of the order list. The known cases where this matters can
be detected by calculating module lengths both for if they are
counted and if they are not. If the latter matches the actual file
length but the former does not, it's most likely one of these edge
case ST modules.